now it compiles for 32 bits

+++ flipper/core/semaphore.c
@@ -14,7 +14,7 @@ FuriSemaphore* furi_semaphore_alloc(uint32_t max_count, uint32_t initial_count) 
     FuriSemaphore* semaphore = malloc(sizeof(FuriSemaphore));
     semaphore->max = max_count;
     semaphore->count = initial_count;
-    semaphore->owners = malloc(sizeof(uint64_t) * max_count);
+    semaphore->owners = malloc(sizeof(size_t) * max_count);
     for(unsigned int i = 0; i < initial_count; i++)
         semaphore->owners[i] = SEMAPHORE_NO_OWNER;
     return semaphore;
@@ -24,7 +24,7 @@ static void* acquire_cb(void* ctx_) {
     FuriSemaphoreCtx* ctx = ctx_;
     while(furi_semaphore_get_count(ctx->semaphore) == ctx->semaphore->max)
         furi_delay_tick(1);
-    ctx->semaphore->owners[ctx->semaphore->count++] = (uint64_t)ctx->id;
+    ctx->semaphore->owners[ctx->semaphore->count++] = (size_t)ctx->id;
     (*ctx->done) = true;
     return NULL;
 }
@@ -59,7 +59,7 @@ FuriStatus furi_semaphore_acquire(FuriSemaphore* instance, uint32_t timeout) {
 FuriStatus furi_semaphore_release(FuriSemaphore* instance) {
     furi_assert(instance);
 
-    uint64_t id = (uint64_t)furi_thread_get_current_id();
+    size_t id = (size_t)furi_thread_get_current_id();
     for(unsigned int i = 0; i < instance->count; i++)
         if(instance->owners[i] == id) {
             memcpy(&instance->owners[i], &instance->owners[i + 1], instance->count - i - 1);
